refactor : Move Project 구조 개선
- 확장 가능성 열어두기 위하여 프로젝트 구성 개선
