+++ src/main/java/com/talearnt/examples/ExampleController.java
@@ -2,10 +2,8 @@ package com.talearnt.examples;
 
 
 import com.talearnt.enums.ErrorCode;
-import com.talearnt.util.CommonResponse;
+import com.talearnt.util.response.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
