+++ src/main/java/com/talearnt/join/JoinController.java
@@ -1,9 +1,9 @@
 package com.talearnt.join;
 
 import com.talearnt.enums.ErrorCode;
-import com.talearnt.exception.CustomException;
-import com.talearnt.util.CommonResponse;
-import com.talearnt.util.RestControllerV1;
+import com.talearnt.util.exception.CustomException;
+import com.talearnt.util.response.CommonResponse;
+import com.talearnt.util.version.RestControllerV1;
 import com.talearnt.verification.VerificationReqDTO;
 import com.talearnt.verification.VerificationService;
 import com.talearnt.verification.VerifyCodeReqDTO;
@@ -12,7 +12,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
